% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_data.R
\name{get_all_data}
\alias{get_all_data}
\title{Get all data from a dataset}
\usage{
get_all_data(
  ds_name,
  base_url_melodi = "https://api.insee.fr/melodi",
  download_directory = tempdir(),
  stringsAsFactors = FALSE,
  remove_file = TRUE
)
}
\arguments{
\item{ds_name}{Dataset name - use get_ds_list to find a dataset identifier}

\item{base_url_melodi}{API Melodi URL - default production URL}

\item{download_directory}{Download directory - default system tempdir}

\item{stringsAsFactors}{if TRUE, strings variables in data.frame are converted to factors - Default FALSE}

\item{remove_file}{Remove downloaded data file after loading data - Default TRUE}
}
\value{
data.frame with data
}
\description{
Get all data from a dataset
}
\examples{
data <- get_all_data("DS_TICM_PRATIQUES")
}
