% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catalog.R
\name{get_catalog}
\alias{get_catalog}
\title{Get list of available datasets}
\usage{
get_catalog(lang = "fr", base_url_melodi = "https://api.insee.fr/melodi")
}
\arguments{
\item{lang}{lang of metadata informations (default: fr)}

\item{base_url_melodi}{API URL (default: https://api.insee.fr)}
}
\value{
a data.frame containing list of available datasets
}
\description{
Get list of available datasets
}
\examples{
get_catalog()
get_catalog(lang = "en")
}
