% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code_list.R
\name{get_code_list}
\alias{get_code_list}
\title{Get a code list by id (id generally starts with 'CL_')}
\usage{
get_code_list(id, lang = "fr")
}
\arguments{
\item{id}{id (id generally starts with 'CL_' : 'CL_AGE', 'CL_LEGAL_FORM', etc.)}

\item{lang}{lang of metadata informations (default: fr)}
}
\value{
a data.frame containing codes and labels
}
\description{
Get a code list by id (id generally starts with 'CL_')
}
\examples{
get_code_list("CL_AGE")
get_code_list(id = "CL_AGE", lang = "en")
}
