% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get Melodi data by URL.
This URL can be copy/paste from explorer https://catalogue-donnees.insee.fr}
\usage{
get_data(url)
}
\arguments{
\item{url}{URL API data Melodi - Start by "https://api.insee.fr/melodi/data/..."}
}
\value{
data.frame with data
}
\description{
Get Melodi data by URL.
This URL can be copy/paste from explorer https://catalogue-donnees.insee.fr
}
\examples{
get_data(
    "https://api.insee.fr/melodi/data/DS_POPULATIONS_REFERENCE?POPREF_MEASURE=PMUN&GEO=FRANCE-F"
)
}
