% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file.R
\name{get_file}
\alias{get_file}
\title{Get file attached to a dataset : extract (XLSX), or full CSV file}
\usage{
get_file(
  ds_name,
  file_name,
  download_file_name,
  download_directory = tempdir(),
  base_url_melodi = "https://api.insee.fr/melodi"
)
}
\arguments{
\item{ds_name}{Dataset name}

\item{file_name}{File name}

\item{download_file_name}{Filename on disk after download}

\item{download_directory}{Download directory - default system tempdir}

\item{base_url_melodi}{API Melodi URL - default production URL}
}
\value{
downloaded file name on disk
}
\description{
Get file attached to a dataset : extract (XLSX), or full CSV file
}
\examples{
get_file(
ds_name = "DS_EC_DECES",
file_name = "T1_DECES_JOUR_NAT_FR",
download_file_name = "T1_DECES_JOUR_NAT_FR.xlsx"
)
}
