% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_local_data.R
\name{get_local_data}
\alias{get_local_data}
\title{Get data on one territory}
\usage{
get_local_data(
  ds_name,
  geo,
  geo_object,
  filter = "",
  base_url_melodi = "https://api.insee.fr/melodi"
)
}
\arguments{
\item{ds_name}{Dataset name}

\item{geo}{Code geo}

\item{geo_object}{Geo Level}

\item{filter}{additionnal filter on request - default : ""}

\item{base_url_melodi}{API Melodi URL - default production URL}
}
\value{
data.frame with data
}
\description{
Get data on one territory
}
\examples{
get_local_data(
  ds_name = "DS_POPULATIONS_REFERENCE",
  geo = "69",
  geo_object = "DEP"
)

get_local_data(
  ds_name = "DS_POPULATIONS_REFERENCE",
  geo = "69",
  geo_object = "DEP",
  filter = "POPREF_MEASURE=PMUN&TIME_PERIOD=2022"
)
}
