% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_range.R
\name{get_range}
\alias{get_range}
\title{Get the dataset's codebook}
\usage{
get_range(
  ds_name,
  base_url_melodi = "https://api.insee.fr/melodi",
  lang = "fr",
  dimension_time_period = TRUE
)
}
\arguments{
\item{ds_name}{dataset name}

\item{base_url_melodi}{API Melodi URL - default production URL}

\item{lang}{french or english labels}

\item{dimension_time_period}{the TIME_PERIOD dimension can include many entries, ignoring it can simplify the results}
}
\value{
A data frame with columns such as \code{dimension}, \code{dimension_label}, \code{value}, \code{value_label}
}
\description{
Retrieves the list of dimensions (except GEO) and all their possible values
(codes and human-readable labels) for a given dataset.
For GEO dimension, refer to get_range_geo()
}
\examples{
get_range("DS_POPULATIONS_REFERENCE")
get_range("DS_RP_EDUCATION",
lang = "en",
dimension_time_period = FALSE)
}
