% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_range_geo.R
\name{get_range_geo}
\alias{get_range_geo}
\title{Get dataset's geography : codes and labels}
\usage{
get_range_geo(
  ds_name,
  base_url_melodi = "https://api.insee.fr/melodi",
  lang = "fr"
)
}
\arguments{
\item{ds_name}{dataset name}

\item{base_url_melodi}{API Melodi URL - default production URL}

\item{lang}{french or english labels - default french ("fr")}
}
\value{
A data frame with columns \code{GEO_REF}, \code{GEO_OBJECT}, \code{GEO}, \code{GEO_LABEL}
}
\description{
Retrieves the list of geo dimensions and all their possible values
(codes and human-readable labels) for a given dataset.
}
\details{
For other dimensions, refer to get_range()
}
\examples{
get_range_geo("DS_POPULATIONS_REFERENCE") |>
head()
}
