% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Get dataset metadata}
\usage{
get_metadata(ds_name, base_url_melodi = "https://api.insee.fr/melodi")
}
\arguments{
\item{ds_name}{Dataset name}

\item{base_url_melodi}{API Melodi URL - default production URL}
}
\value{
list dataset metadata
}
\description{
Get dataset metadata
}
\examples{
get_metadata("DS_POPULATIONS_REFERENCE")
}
