% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_range.R
\name{get_range}
\alias{get_range}
\title{Get dataset range (codes and labels)}
\usage{
get_range(
  ds_name,
  base_url_melodi = "https://api.insee.fr/melodi",
  lang = "fr",
  exclusions_list = c("GEO", "TIME_PERIOD")
)
}
\arguments{
\item{ds_name}{dataset name}

\item{base_url_melodi}{API Melodi URL - default production URL}

\item{lang}{french or english labels - default french}

\item{exclusions_list}{exclude some dimensions for a faster and light result - default : "GEO", "TIME_PERIOD"}
}
\value{
A data frame with dimensions and modalities codes and labels
}
\description{
Retrieves the list of dimensions and all their possible modalities values
(codes and human-readable labels) for a given dataset.
For GEO dimension, it is recommanded to used dedicated function : get_range_geo()
}
\examples{
get_range("DS_POPULATIONS_REFERENCE")
get_range(ds_name = "DS_EC_DECES", lang = "en")
}
